# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pakperk API",
    "description": "Stable v1 paper-reading API with optional OIDC-backed account profiles. Library and comment routes remain feature-gated until their production phases are complete.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/papers/{paper_id}/chat": {
      "post": {
        "tags": [
          "routes::chat"
        ],
        "operationId": "chat",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Grounded chat answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Invalid anonymous session or question",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
...
    "/v1/papers/{paper_id}/processing": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "processing",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Processing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStateSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CapabilitiesSchema": {
        "type": "object",
        "required": [
          "metadata",
          "introduction",
          "chat",
          "connections"
        ],
        "properties": {
          "chat": {
            "type": "boolean"
          },
          "connections": {
            "type": "boolean"
          },
          "introduction": {
            "type": "boolean"
          },
          "metadata": {
            "type": "boolean"
          }
        }
      },
      "ChatBody": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
      "ChatEvidenceSchema": {
        "type": "object",
        "required": [
          "section_kind",
          "chunk_id"
        ],
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid"
          },
          "page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "section_heading": {
            "type": [
              "string",
              "null"
            ]
          },
          "section_kind": {
            "$ref": "#/components/schemas/SectionKindSchema"
          }
        }
      },
      "ChatResponseSchema": {
        "type": "object",
        "required": [
          "thread_id",
          "generation",
          "answer_markdown",
          "insufficient_evidence",
          "evidence",
          "suggested_follow_ups",
          "prompt_version"
        ],
        "properties": {
          "answer_markdown": {
            "type": "string"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatEvidenceSchema"
            }
          },
          "generation": {
            "type": "integer",
            "format": "int32"
          },
          "insufficient_evidence": {
            "type": "boolean"
          },
          "model_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "prompt_version": {
            "type": "string"
          },
          "provider_request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "suggested_follow_ups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thread_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ErrorBodySchema": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code.",
            "example": "PAPER_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "The requested paper was not found."
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ErrorEnvelopeSchema": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodySchema"
          }
        }
      },
      "FailureCategorySchema": {
        "type": "string",
        "enum": [
          "external_temporary",
          "external_permanent",
          "parser_temporary",
          "parser_document",
          "model_temporary",
          "validation",
          "internal"
        ]
      },
...
      "OverallProcessingStateSchema": {
        "type": "string",
        "enum": [
          "not_requested",
          "processing",
          "ready",
          "failed"
        ]
      },
...
      "ProcessingErrorSchema": {
        "type": "object",
        "required": [
          "category",
          "code",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FailureCategorySchema"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ProcessingStageSchema": {
        "type": "string",
        "enum": [
          "not_requested",
          "queued",
          "fetching_license",
          "fetching_pdf",
          "parsing_pdf",
          "introduction_ready",
          "indexing_chat",
          "resolving_references",
          "ready",
          "failed_retryable",
          "failed_terminal"
        ]
      },
      "ProcessingStateSchema": {
        "type": "object",
        "required": [
          "paper_id",
          "generation",
          "overall_state",
          "stage",
          "capabilities",
          "retryable",
          "updated_at"
        ],
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesSchema"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "embedding_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "generation": {
            "type": "integer",
            "format": "int32"
          },
          "last_error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessingErrorSchema"
              }
            ]
          },
          "overall_state": {
            "$ref": "#/components/schemas/OverallProcessingStateSchema"
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "parser_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "retryable": {
            "type": "boolean"
          },
          "stage": {
            "$ref": "#/components/schemas/ProcessingStageSchema"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "SectionKindSchema": {
        "type": "string",
        "enum": [
          "abstract",
          "introduction",
          "background",
          "related_work",
          "method",
          "experiment",
          "result",
          "discussion",
          "limitation",
          "conclusion",
          "appendix",
          "acknowledgment",
          "references",
          "other"
        ]
      }
    },
    "securitySchemes": {
      "oidcBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC access token validated against the configured issuer, audience, algorithm, and JWKS."
      }
    }
  },
...
}